feat: tag delete with hook
